+++ packages/side-eyes/src/app/components/MoreInfo/index.jsx
@@ -4,11 +4,7 @@ import Link from '../../../commons/components/Link'
 export default function MoreInfo() {
   return (
     <span>
-      For more information read the{' '}
-      <Link href="https://applitools.com/tutorials/selenium-ide.html">
-        tutorial
-      </Link>
-      .
+      For more information read the <Link href="https://applitools.com/tutorials/selenium-ide.html">tutorial</Link>.
     </span>
   )
 }