+++ packages/side-eyes/src/background/__test__/external-state.spec.js
@@ -1,9 +1,4 @@
-import {
-  getExternalState,
-  setExternalState,
-  setExternalStateInternally,
-  resetMode,
-} from '../external-state'
+import { getExternalState, setExternalState, setExternalStateInternally, resetMode } from '../external-state'
 jest.mock('webextension-polyfill')
 
 describe('external-state', () => {