+++ packages/side-eyes/src/commons/components/Combobox/index.jsx
@@ -36,9 +36,7 @@ export default class Combobox extends React.Component {
     this.setState({ isOpen: false })
   }
   calculateModalPosition() {
-    const rect = this.button
-      ? this.button.getBoundingClientRect()
-      : { top: 0, left: 0 }
+    const rect = this.button ? this.button.getBoundingClientRect() : { top: 0, left: 0 }
     return {
       top: `${rect.top}px`,
       left: `${rect.left}px`,
@@ -57,11 +55,7 @@ export default class Combobox extends React.Component {
         }}
         className={this.props.className}
       >
-        {this.props.label && (
-          <div>
-            {this.props.label}
-          </div>
-        )}
+        {this.props.label && <div>{this.props.label}</div>}
         <a
           className={classNames('combobox', { disabled: this.props.disabled })}
           ref={button => {
@@ -71,9 +65,7 @@ export default class Combobox extends React.Component {
           onClick={this.openCombobox.bind(this)}
         >
           <React.Fragment>
-            <span style={{ marginRight: '5px' }}>
-              {this.props.items[this.state.selectedIndex]}
-            </span>
+            <span style={{ marginRight: '5px' }}>{this.props.items[this.state.selectedIndex]}</span>
             <ArrowIndicator directionIsUp={this.state.isOpen} />
           </React.Fragment>
         </a>
@@ -86,17 +78,8 @@ export default class Combobox extends React.Component {
         >
           <ul className="combobox-content">
             {this.props.items.map((item, index) => (
-              <li
-                key={item}
-                name={item}
-                label={this.props.label}
-                onClick={this.props.onClick}
-              >
-                <a
-                  className="combobox-item"
-                  href="#"
-                  onClick={() => this.click(item, index)}
-                >
+              <li key={item} name={item} label={this.props.label} onClick={this.props.onClick}>
+                <a className="combobox-item" href="#" onClick={() => this.click(item, index)}>
                   {item}
                 </a>
               </li>