+++ packages/visual-grid-client/CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## Unreleased
 
+
+ ## 14.4.4 - 2020/6/9
+
 - added chromeEmulationInfo and TS defs for iosDeviceInfo 
+- updated to @applitools/dom-snapshot@3.5.3 (from 3.5.2)
+- updated to @applitools/eyes-sdk-core@11.0.5 (from 11.0.4)
 
 ## 14.4.3 - 2020/6/4